+++ src/app/core/auth/http-interceptors.ts
@@ -1,5 +1,4 @@
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
-import { TokenCheckInterceptor } from '../interceptors/token-check.interceptor';
 import { AuthInterceptorService } from 'src/app/core/interceptors/auth-interceptor.service';
 
 export const httpInterceptors = [
@@ -9,16 +8,3 @@ export const httpInterceptors = [
     multi: true,
   },
 ];
-
-// export const httpInterceptors = [
-//   {
-//     provide: HTTP_INTERCEPTORS,
-//     useClass: TokenCheckInterceptor,
-//     multi: true,
-//   },
-//   {
-//     provide: HTTP_INTERCEPTORS,
-//     useClass: AuthInterceptorService,
-//     multi: true,
-//   },
-// ];